requirements.txt
numpy>=1.24
scipy>=1.10
pandas>=2.0
matplotlib>=3.7
ipykernel>=6.0
jupyter>=1.0
openai>=1.30
anthropic>=0.34
google-generativeai>=0.7
